Format strings and refactor user input handling
Updated Messages.cs to add color formatting to the "Back" option and centralized user input handling through UserPrompt in command classes. Refactored various commands to remove redundant methods and streamline user choices and error handling, improving code readability and maintainability.
