+++ Drinks.Eddyfadeev/Drinks.Eddyfadeev/View/Messages.cs
@@ -4,7 +4,7 @@ internal static class Messages
 {
     internal const string PressEnterToContinue = "[green]Press Enter to continue...[/]";
     internal const string NoDrinksFound = "[red]No drinks found.[/]";
-    internal const string BackOption = "Back";
+    internal const string BackOption = "[yellow]Back[/]";
     internal const string ApiConfigurationError = "[red]API configuration error. Please check the configuration file.[/]";
     internal const string ExitingApplication = "[white]Exiting application...[/]";
     internal const string ReturningToMainMenu = "[white]Returning to the main menu...[/]";
